Added too many things to list. See description
Reorganized things into separate namespaces within Stater, also made folders corresponding to them.

Built the first implementation of the core assembly analyzer. Has a "pass" system reminicent of LLVM, but obviously less fancy. These passes are defined by implementing `IAnalysisPass<T>` where `T` is what it is analysing (see the code for what those are). To accompany that there are two passes added so far: one which simply adds each class as a node, and another which finds all usages of the Unity Input class and creates an incoming edge to the node representing it.

On that note, there is now a constant representing an external input (like the Unity Input class) which lives in StaterConstants.cs. We could also move to having one constant per external input, but I think that's a bit too rigid. It makes a bit more sense to me that we'd have one psudo-node representing external input and the data annotated on the edge gives the information about it.

I added a new Stater.Utils namespace which currently holds a single class named `t`. This class lives in `TextUtils.cs`, so it should hopefully be obvious what the `t` stands for: `TextUtils`. The point of the class was to wrap common but custom text and/or output operations in a simple class that can be easily referenced, hence the single letter name. The two functions in it are `tab` which prints spaces corresponding to a tab level supplied as input, and `debugPrint` which takes a main message as well as a value to "enclose" (or wrap in quotes) as I like to use for debug.

To accompany these text utilities (and in fact inspiring them) is the addition of `ToString()` methods on `Node` and `ClassGraph`. These should help debugging efforts.

Lastly, there's some "legacy" code I wrote to detect usages of tags, back before I added the pass system. I'm leaving this in as an example of how to get that information, in case we ever want it again.